+++ History.md
@@ -5,6 +5,7 @@ unreleased
   * Properly support all HTTP methods
   * Support vanilla node.js http servers
   * Treat `ENAMETOOLONG` as code 414
+  * Use accepts for negotiation
 
 1.0.3 / 2014-05-20
 ==================